pyproject.toml
[tool.poetry]
name = "brhaneblog"
version = "0.1.0"
description = "This is my blog created with django"
authors = ["eri-bga <godseeshears@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
django-tinymce = "^3.5.0"
fontawesomefree = "^6.2.1"
django-recaptcha = "^3.0.0"
psycopg2-binary = "^2.9.5"
django-allauth = "^0.52.0"
django-taggit = "^3.1.0"
markdown = "^3.4.1"
django-tailwind = "^3.4.0"
django-widget-tweaks = "^1.4.12"
pillow = "^9.4.0"
django-crispy-forms = "^1.14.0"
six = "^1.16.0"
django-debug-toolbar = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
